refactor: place saveAll in new class hiearchy
Required to avoid breaking changes in client code that uses monguito with standalone Mongo.
