+++ src/index.ts
@@ -4,7 +4,9 @@ import {
   TypeData,
   TypeMap,
 } from './mongoose.repository';
+import { MongooseTransactionalRepository } from './mongoose.transactional-repository';
 import { PartialEntityWithId, Repository } from './repository';
+import { TransactionalRepository } from './transactional-repository';
 import { Auditable, AuditableClass, isAuditable } from './util/audit';
 import { Entity } from './util/entity';
 import {
@@ -25,8 +27,10 @@ export {
   Entity,
   IllegalArgumentException,
   MongooseRepository,
+  MongooseTransactionalRepository,
   PartialEntityWithId,
   Repository,
+  TransactionalRepository,
   TypeData,
   TypeMap,
   UndefinedConstructorException,
